requirements.txt
asteval==0.9.29
asttokens==2.4.1
backports.tarfile==1.2.0
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
configparser==6.0.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
docutils==0.21.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.39.3
future==0.18.3
idna==3.7
importlib-resources==5.12.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.2.1
kiwisolver==1.4.4
laserbeamsize==1.9.4
lmfit==1.1.0
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
nest-asyncio==1.6.0
nh3==0.2.17
numpy==1.24.2
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
parso==0.8.4
pathlib==1.0.1
Pillow==9.5.0
pkginfo==1.11.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-can==4.3.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
readme_renderer==43.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
scipy==1.10.1
shutils==0.1.0
six==1.16.0
stack-data==0.6.3
tk==0.1.0
tornado==6.4
traitlets==5.14.3
twine==5.1.0
typing_extensions==4.12.1
tzdata==2023.3
uncertainties==3.1.7
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.15.0
uptime==3.0.1
pyserial==3.5